Changes after setting up new un-forked repo.

+++ data/tests/bragging-hunter/test.sh
@@ -29,8 +29,9 @@ printf "Using checkpoint: $cp\n"
 tmp="$z/tmp.txt"
 cp /dev/null "$tmp"
 
+v=("02-ru" "03-ru" "01-chr" "02-chr" "03-chr" "04-chr" "05-chr")
 #v=("02-ru" "03-ru" "01-chr" "02-chr" "03-chr" "04-chr")
-v=("14-de" "51-de" "02-fr" "04-fr" "14-fr" "18-fr" "19-fr" "22-fr" "03-ru" "03-chr")
+#v=("14-de" "51-de" "02-fr" "04-fr" "14-fr" "18-fr" "19-fr" "22-fr" "03-ru" "03-chr")
 vsize="${#v[@]}"
 
 printf "\nTotal voice count: %d\n\n" "$vsize"
@@ -43,7 +44,8 @@ for voice in "${v[@]}"; do
 	printf "Generating audio for %s\n" "$voice"
 	ix=0
 	syn=""
-	cat "$text" | while read sentence; do
+	cp /dev/null "$tmp"
+	cat "$text" | head -n 1 | while read sentence; do
 		ix=$(($ix+1))
 		printf "%d|%s|%s|chr\n" "$ix" "${sentence}" "$voice" >> "$tmp"
 	done
@@ -61,13 +63,14 @@ for voice in "${v[@]}"; do
 	#python wavernnx-cpu.py
 	python wavernnx.py
 	
-	count=$(wc -l "$selected"|cut -f 1 -d ' ')
+	count=$(wc -l "$text"|cut -f 1 -d ' ')
 	for ix in $(seq 1 $count); do
 		iy=$(printf "%02d" $ix)
-		wav="$wg"-"$voice/wg-$ix.wav"
-		mv "$ix.wav" "$wav"
+		wav="$wg"-"$voice/$voice-$wg-$ix.wav"
+		if [ ! -f "wg-$ix.wav" ]; then continue; fi
+		mv "wg-$ix.wav" "$wav"
 		mp3="$wg"-"$voice/$wg-$voice-$iy".mp3
-		ffmpeg -i "$wav" -codec:a libmp3lame -qscale:a 4 "$mp3" 2>&1 > /dev/null
+		ffmpeg -i "$wav" -codec:a libmp3lame -qscale:a 4 "$mp3" > /dev/null 2>&1
 		rm "$wav"
 	done
 
@@ -76,9 +79,8 @@ for voice in "${v[@]}"; do
 	count=$(wc -l "$text"|cut -f 1 -d ' ')
 	for ix in $(seq 1 $count); do
 		ix="$(($ix+1))"
-		rm "$ix".wav
-		rm "wg-$ix".wav
-		rm "$ix".npy
+		rm "$ix".wav 2> /dev/null || true
+		rm "$ix".npy 2> /dev/null || true
 	done
 	printf "\n\n"
 done
